Ensure that RTC read is valid.
Stm32 RTC shadow registers are not guarenteed to be valid after CPU has
been stopped. As RTC is used in interrupt, it is not possible to wait
for them to be updated. Instead, bypass the shadow registers and read
RTC directly, using algorithm instructed by reference manual (read two
times, maybe three).